+++ runtime/bin/file.cc
@@ -763,7 +763,7 @@ void FUNCTION_NAME(File_AreIdentical)(Dart_NativeArguments args) {
   Namespace* namespc = Namespace::GetNamespace(args, 0);
   const char* path_1 = DartUtils::GetNativeStringArgument(args, 1);
   const char* path_2 = DartUtils::GetNativeStringArgument(args, 2);
-  File::Identical result = File::AreIdentical(namespc, path_1, path_2);
+  File::Identical result = File::AreIdentical(namespc, path_1, namespc, path_2);
   if (result == File::kError) {
     Dart_SetReturnValue(args, DartUtils::NewDartOSError());
   } else {
@@ -1455,7 +1455,7 @@ CObject* File::IdenticalRequest(const CObjectArray& request) {
   CObjectString path1(request[1]);
   CObjectString path2(request[2]);
   File::Identical result =
-      File::AreIdentical(namespc, path1.CString(), path2.CString());
+      File::AreIdentical(namespc, path1.CString(), namespc, path2.CString());
   if (result == File::kError) {
     return CObject::NewOSError();
   }