+++ sdk/lib/io/socket.dart
@@ -8,11 +8,15 @@ part of dart.io;
 
 /**
  * [InternetAddressType] is the type an [InternetAddress]. Currently,
- * IP version 4 (IPv4) and IP version 6 (IPv6) are supported.
+ * IP version 4 (IPv4), IP version 6 (IPv6) and Unix domain address are
+ * supported. Unix domain sockets are available only on Linux, MacOS and
+ * Android.
  */
 class InternetAddressType {
   static const InternetAddressType IPv4 = const InternetAddressType._(0);
   static const InternetAddressType IPv6 = const InternetAddressType._(1);
+  @Since("2.8")
+  static const InternetAddressType unix = const InternetAddressType._(2);
   static const InternetAddressType any = const InternetAddressType._(-1);
 
   @Deprecated("Use IPv4 instead")
@@ -27,8 +31,9 @@ class InternetAddressType {
   const InternetAddressType._(this._value);
 
   factory InternetAddressType._from(int value) {
-    if (value == 0) return IPv4;
-    if (value == 1) return IPv6;
+    if (value == IPv4._value) return IPv4;
+    if (value == IPv6._value) return IPv6;
+    if (value == unix._value) return unix;
     throw new ArgumentError("Invalid type: $value");
   }
 
@@ -43,6 +48,8 @@ class InternetAddressType {
         return "IPv4";
       case 1:
         return "IPv6";
+      case 2:
+        return "Unix";
       default:
         throw new ArgumentError("Invalid InternetAddress");
     }
@@ -52,7 +59,7 @@ class InternetAddressType {
 }
 
 /**
- * An internet address.
+ * An internet address or a Unix domain address.
  *
  * This object holds an internet address. If this internet address
  * is the result of a DNS lookup, the address also holds the hostname
@@ -95,27 +102,35 @@ abstract class InternetAddress {
   external static InternetAddress get ANY_IP_V6;
 
   /**
-   * The [type] of the [InternetAddress] specified what IP protocol.
+   * The address family of the [InternetAddress].
    */
   InternetAddressType get type;
 
   /**
-   * The numeric address of the host. For IPv4 addresses this is using
-   * the dotted-decimal notation. For IPv6 it is using the
-   * hexadecimal representation.
+   * The numeric address of the host.
+   *
+   * For IPv4 addresses this is using the dotted-decimal notation.
+   * For IPv6 it is using the hexadecimal representation.
+   * For Unix domain addresses, this is a file path.
    */
   String get address;
 
   /**
-   * The host used to lookup the address. If there is no host
-   * associated with the address this returns the numeric address.
+   * The host used to lookup the address.
+   *
+   * If there is no host associated with the address this returns the [address].
    */
   String get host;
 
   /**
-   * Get the raw address of this [InternetAddress]. The result is either a
-   * 4 or 16 byte long list. The returned list is a copy, making it possible
-   * to change the list without modifying the [InternetAddress].
+   * The raw address of this [InternetAddress].
+   *
+   * For an IP address, the result is either a 4 or 16 byte long list.
+   * For a Unix domain address, UTF-8 encoded byte sequences that represents
+   * [address] is returned.
+   *
+   * The returned list is a fresh copy, making it possible to change the list without
+   * modifying the [InternetAddress].
    */
   Uint8List get rawAddress;
 
@@ -135,17 +150,48 @@ abstract class InternetAddress {
   bool get isMulticast;
 
   /**
-   * Creates a new [InternetAddress] from a numeric address.
+   * Creates a new [InternetAddress] from a numeric address or a file path.
    *
-   * If the address in [address] is not a numeric IPv4
-   * (dotted-decimal notation) or IPv6 (hexadecimal representation).
-   * address [ArgumentError] is thrown.
+   * If [type] is [InternetAddressType.IPv4], [address] must be a numeric IPv4
+   * address (dotted-decimal notation).
+   * If [type] is [InternetAddressType.IPv6], [address] must be a numeric IPv6
+   * address (hexadecimal notation).
+   * If [type] is [InternetAddressType.unix], [address] must be a a valid file
+   * path.
+   * If [type] is omitted, [address] must be either a numeric IPv4 or IPv6
+   * address and the type is inferred from the format.
+   *
+   * To create a Unix domain address, [type] should be
+   * [InternetAddressType.unix] and [address] should be a string.
    */
-  external factory InternetAddress(String address);
+  external factory InternetAddress(String address,
+      {@Since("2.8") InternetAddressType type});
 
   /**
-   * Perform a reverse dns lookup on the [address], creating a new
-   * [InternetAddress] where the host field set to the result.
+   * Creates a new [InternetAddress] from the provided raw address bytes.
+   *
+   * If the [type] is [InternetAddressType.IPv4], the [rawAddress] must have
+   * length 4.
+   * If the [type] is [InternetAddressType.IPv6], the [rawAddress] must have
+   * length 16.
+   * If the [type] is [InternetAddressType.IPv4], the [rawAddress] must be a
+   * valid UTF-8 encoded file path.
+   *
+   * If [type] is omitted, the [rawAddress] must have a length of either 4 or
+   * 16, in which case the type defaults to [InternetAddress.IPv4] or
+   * [InternetAddress.IPv6] respectively.
+   */
+  external factory InternetAddress.fromRawAddress(Uint8List rawAddress,
+      {@Since("2.8") InternetAddressType type});
+
+  /**
+   * Perform a reverse DNS lookup on this [address]
+   *
+   * Returns a new [InternetAddress] with the same address, but where the [host]
+   * field set to the result of the lookup.
+   *
+   * If this address is Unix domain addresses, no lookup is performed and this
+   * address is returned directly.
    */
   Future<InternetAddress> reverse();
 
@@ -826,28 +872,30 @@ abstract class Socket implements Stream<Uint8List>, IOSink {
   void setRawOption(RawSocketOption option);
 
   /**
-   * Returns the port used by this socket.
+   * The port used by this socket.
    *
    * Throws a [SocketException] if the socket is closed.
+   * The port is 0 if the socket is a Unix domain socket.
    */
   int get port;
 
   /**
-   * Returns the remote port connected to by this socket.
+   * The remote port connected to by this socket.
    *
    * Throws a [SocketException] if the socket is closed.
+   * The port is 0 if the socket is a Unix domain socket.
    */
   int get remotePort;
 
   /**
-   * Returns the [InternetAddress] used to connect this socket.
+   * The [InternetAddress] used to connect this socket.
    *
    * Throws a [SocketException] if the socket is closed.
    */
   InternetAddress get address;
 
   /**
-   * Returns the remote [InternetAddress] connected to by this socket.
+   * The remote [InternetAddress] connected to by this socket.
    *
    * Throws a [SocketException] if the socket is closed.
    */