+++ tests/standalone_2/io/socket_info_ipv6_test.dart
@@ -2,8 +2,11 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-import "package:expect/expect.dart";
+import "dart:async";
 import "dart:io";
+import "dart:typed_data";
+
+import "package:expect/expect.dart";
 
 void testHostAndPort() {
   ServerSocket.bind("::1", 0).then((server) {
@@ -29,6 +32,30 @@ void testHostAndPort() {
   });
 }
 
-void main() {
+Future<void> testRawAddress() async {
+  var list =
+      Uint8List.fromList([0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1]);
+  var addr = '::1';
+  var address = InternetAddress.fromRawAddress(list);
+  Expect.equals(address.address, addr);
+  var server = await ServerSocket.bind(address, 0);
+  var client = await Socket.connect(address, server.port);
+  var completer = Completer<void>();
+  server.listen((socket) async {
+    Expect.equals(socket.port, server.port);
+    Expect.equals(client.port, socket.remotePort);
+    Expect.equals(client.remotePort, socket.port);
+
+    Expect.equals(client.remoteAddress, address);
+    socket.destroy();
+    client.destroy();
+    await server.close();
+    completer.complete();
+  });
+  await completer.future;
+}
+
+void main() async {
   testHostAndPort();
+  await testRawAddress();
 }